feat: add SpellCircleOverlay component with animated spell circles and feedback effects
